// src/app/types/user.ts
export interface User {
  avatar: string;
  email: string;
  firstName: string;
  follow: boolean;
  id: string;
  lastName: string;
}
export interface UserInfo {
  birthday: string;
  city: string;
  work: string; //плохое название
  familyStatus: string;
  nik: string; //плохое название
  status: string;
  id: string;
  firstName: string;
  lastName: string;
  avatar: string;
}