// src/app/types/profile.ts
export interface ProfileInfo {
  dateOfBirth: string;
  city: string;
  placeOfWork: string;
  familyStatus: string;
  nik: string;
  status: string;
}
export interface Profile {
  id: string;
  firstName: string;
  lastName: string;
  avatar: string;
  follow: boolean;
  email: string;
  token: string;
  password: string;
}